requirements.txt
Django==1.5.1
South==0.7.6
boto==2.9.0
dj-database-url==0.2.1
django-storages==1.1.8
gunicorn==0.17.2
psycopg2==2.5
wsgiref==0.1.2
